pyproject.toml
[tool.poetry]
name = "nonebot-plugin-xjie-weather"
version = "1.2.2"

description = "A little weather plugin"
authors = ["AwAjie <139576615+ajdgg@users.noreply.github.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/ajdgg/nonebot-plugin-xjie-weather"
repository = "https://github.com/ajdgg/nonebot-plugin-xjie-weather"
documentation = "https://github.com/ajdgg/nonebot-plugin-xjie-weather"
keywords = ["nonebot", "nonebot2"]


[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.2.0"
Jinja2 = "^3.1.4"
playwright = "^1.44.0"
httpx = "^0.27.0"
Pillow = "^10.4.0"
nonebot-plugin-alconna = "^0.46.4"
nonebot-adapter-onebot = { version = "^2.0.0-beta.1", optional = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
